apparently working with odyssey 1.1
